Reorganize db schema files
