+++ src/server/db/schema/images.ts
@@ -6,11 +6,11 @@ import {
   text,
   tinytext,
 } from 'drizzle-orm/mysql-core'
-import { s3ObjectKey } from '../utilities'
+import { dbS3ObjectKey } from '../utilities'
 
 export const images = mysqlTable('image', {
   id: serial('id').primaryKey(),
-  key: s3ObjectKey('key').notNull(),
+  key: dbS3ObjectKey('key').notNull(),
   width: int('width').notNull(),
   height: int('height').notNull(),
   source: text('source'),
