+++ src/server/db/schema/index.ts
@@ -2,6 +2,7 @@ export * from './auth'
 export * from './externalLinks'
 export * from './features'
 export * from './images'
+export * from './placeCategories'
 export * from './placeLists'
 export * from './places'
 export * from './routes'
