+++ src/server/db/schema/routes.ts
@@ -1,6 +1,5 @@
 import { relations } from 'drizzle-orm'
 import {
-  boolean,
   double,
   int,
   mysqlTable,
@@ -10,12 +9,11 @@ import {
 } from 'drizzle-orm/mysql-core'
 import { multiLineType } from '~/server/helpers/spatial-data/multi-line'
 import { mysqlTableWithTranslations } from '../../helpers/translations/db-tables'
-import { colorNames, iconNames } from '../constants/shared'
-import { gender } from '../utilities'
 import { externalLinks } from './externalLinks'
 import { features } from './features'
 import { images } from './images'
 import { places } from './places'
+import { routeCategories, routesToRouteCategories } from './routeCategories'
 
 export const {
   normalTable: routes,
@@ -61,65 +59,6 @@ export const routesRelations = relations(routes, (r) => ({
     references: [features.id],
   }),
 }))
-
-export const {
-  normalTable: routeCategories,
-  translationsTable: routeCategoriesTranslations,
-  makeRelationsWithTranslations: makeRouteCategoryRelations,
-  translationsTableRelations: routeCategoriesTranslationsRelations,
-} = mysqlTableWithTranslations({
-  name: 'routeCategory',
-  normalColumns: {
-    icon: tinytext('icon', { enum: iconNames }).notNull(),
-    color: tinytext('color', { enum: colorNames }).notNull(),
-    hasVisitMission: boolean('hasVisitMission').notNull().default(true),
-    order: int('order'),
-  },
-  translatableColumns: {
-    name: tinytext('name').notNull(),
-    namePlural: tinytext('namePlural').notNull(),
-    nameGender: gender('nameGender'),
-  },
-})
-
-export const routeCategoriesRelations = relations(routeCategories, (r) => ({
-  ...makeRouteCategoryRelations(r),
-
-  mainRoutes: r.many(routes, { relationName: 'mainCategory' }),
-  routes: r.many(routesToRouteCategories, {
-    relationName: 'secondaryCategories',
-  }),
-}))
-
-export const routesToRouteCategories = mysqlTable(
-  'routeToRouteCategory',
-  {
-    routeId: int('routeId').notNull(),
-    categoryId: int('categoryId').notNull(),
-  },
-  (table) => {
-    return {
-      pk: primaryKey({
-        columns: [table.categoryId, table.routeId],
-      }),
-    }
-  }
-)
-
-export const routesToRouteCategoriesRelations = relations(
-  routesToRouteCategories,
-  ({ one }) => ({
-    route: one(routes, {
-      fields: [routesToRouteCategories.routeId],
-      references: [routes.id],
-      relationName: 'secondaryCategories',
-    }),
-    category: one(routeCategories, {
-      fields: [routesToRouteCategories.categoryId],
-      references: [routeCategories.id],
-    }),
-  })
-)
 
 export const routesToPlaces = mysqlTable(
   'routeToPlace',
